requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
